requirements.txt
-i https://pypi.org/simple
alembic==1.0.9
amqp==2.4.2
aniso8601==6.0.0
attrs==19.1.0
backcall==0.1.0
billiard==3.6.0.0
bleach==3.1.0
cachetools==3.1.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
click==7.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dynaconf[yaml]==2.0.1
entrypoints==0.3
flask-caching==1.7.1
flask-jwt-extended==3.18.1
flask-marshmallow==0.10.0
flask-migrate==2.4.0
flask-restplus==0.12.1
flask-sqlalchemy==2.3.2
flask==1.0.2
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
google-auth==1.6.3
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
ipykernel==5.1.0
ipython-genutils==0.2.0
ipython==7.4.0 ; python_version >= '3.3'
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter==1.0.0
kiwisolver==1.0.1
kombu==4.5.0
mako==1.0.9
markupsafe==1.1.1
marshmallow-sqlalchemy==0.16.2
marshmallow==2.19.2
matplotlib==3.0.3
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
oauthlib==3.0.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0 ; sys_platform != 'win32'
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0 ; os_name != 'nt'
pyasn1-modules==0.2.4
pyasn1==0.4.5
pygments==2.3.1
pyjwt==1.7.1
pymysql==0.9.3
pyparsing==2.4.0
pyrsistent==0.14.11
python-box==3.3.0
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.1
pyyaml==5.1
pyzmq==18.0.1
qtconsole==4.4.3
redis==3.2.1
requests-oauthlib==1.2.0
requests==2.21.0
rsa==4.0
send2trash==1.5.0
six==1.12.0
sqlalchemy==1.3.3
terminado==0.8.2
testpath==0.4.2
toml==0.10.0
tornado==6.0.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.2
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
werkzeug==0.15.2
widgetsnbextension==3.4.2
